requirements.txt
   python-binance==1.0.17
   python-telegram-bot==20.3
   python-dotenv==1.0.0
   requests==2.31.0